[SYCLomatic][NFC] Move rules for Math/driver/runtime API to RulesLang folder.
