+++ clang/lib/DPCT/RulesLang/APINamesAtomic.inc
