+++ clang/lib/DPCT/RulesLang/APINamesCooperativeGroups.inc
