+++ clang/lib/DPCT/RulesLang/APINamesErrorHandling.inc
