+++ clang/lib/DPCT/RulesLang/APINamesEvent.inc
