+++ clang/lib/DPCT/RulesLang/APINamesGraphicsInterop.inc
