+++ clang/lib/DPCT/RulesLang/APINamesMemory.inc
