+++ clang/lib/DPCT/RulesLang/APINamesMisc.inc
