+++ clang/lib/DPCT/RulesLang/APINamesStream.inc
