+++ clang/lib/DPCT/RulesLang/APINamesTexture.inc
