+++ clang/lib/DPCT/RulesLang/APINamesWarp.inc
