+++ clang/lib/DPCT/RulesLang/APINamesWmma.inc
