+++ clang/lib/DPCT/RulesLang/CallExprRewriterAtomic.cpp
@@ -1,4 +1,4 @@
-//===--------------- CallExprRewriterWarp.cpp -----------------------------===//
+//===--------------- CallExprRewriterAtomic.cpp ---------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -11,14 +11,14 @@
 
 namespace clang {
 namespace dpct {
-
-void CallExprRewriterFactoryBase::initRewriterMapWarp() {
+// clang-format off
+void CallExprRewriterFactoryBase::initRewriterMapAtomic() {
   RewriterMap->merge(
       std::unordered_map<std::string,
                          std::shared_ptr<CallExprRewriterFactoryBase>>({
-
-#include "APINamesWarp.inc"
+#include "RulesLang/APINamesAtomic.inc"
       }));
+      // clang-format on
 }
 
 } // namespace dpct
