+++ clang/lib/DPCT/RulesLang/CallExprRewriterComplex.cpp
@@ -1,4 +1,4 @@
-//===----------------- CallExprRewriterGraphicsInterop.cpp ----------------===//
+//===--------------- CallExprRewriterComplex.cpp --------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -12,11 +12,11 @@
 namespace clang {
 namespace dpct {
 
-void CallExprRewriterFactoryBase::initRewriterMapGraphicsInterop() {
+void CallExprRewriterFactoryBase::initRewriterMapComplex() {
   RewriterMap->merge(
       std::unordered_map<std::string,
                          std::shared_ptr<CallExprRewriterFactoryBase>>({
-#include "APINamesGraphicsInterop.inc"
+#include "RulesLang/APINamesComplex.inc"
       }));
 }
