+++ clang/lib/DPCT/RulesLang/CallExprRewriterDriver.cpp
@@ -52,7 +52,7 @@ void CallExprRewriterFactoryBase::initRewriterMapDriver() {
   RewriterMap->merge(
       std::unordered_map<std::string,
                          std::shared_ptr<CallExprRewriterFactoryBase>>({
-#include "APINamesDriver.inc"
+#include "RulesLang/APINamesDriver.inc"
       }));
 }
