+++ clang/lib/DPCT/RulesLang/CallExprRewriterErrorHandling.cpp
@@ -29,7 +29,7 @@ void CallExprRewriterFactoryBase::initRewriterMapErrorHandling() {
   RewriterMap->merge(
       std::unordered_map<std::string,
                          std::shared_ptr<CallExprRewriterFactoryBase>>({
-#include "APINamesErrorHandling.inc"
+#include "RulesLang/APINamesErrorHandling.inc"
       }));
 }
