+++ clang/lib/DPCT/RulesLang/CallExprRewriterGraphicsInterop.cpp
@@ -1,4 +1,4 @@
-//===--------------- CallExprRewriterGraph.cpp ----------------------------===//
+//===----------------- CallExprRewriterGraphicsInterop.cpp ----------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -12,11 +12,11 @@
 namespace clang {
 namespace dpct {
 
-void CallExprRewriterFactoryBase::initRewriterMapGraph() {
+void CallExprRewriterFactoryBase::initRewriterMapGraphicsInterop() {
   RewriterMap->merge(
       std::unordered_map<std::string,
                          std::shared_ptr<CallExprRewriterFactoryBase>>({
-#include "APINamesGraph.inc"
+#include "RulesLang/APINamesGraphicsInterop.inc"
       }));
 }
