+++ clang/lib/DPCT/RulesLang/CallExprRewriterTexture.cpp
@@ -260,7 +260,7 @@ void CallExprRewriterFactoryBase::initRewriterMapTexture() {
   BIND_TEXTURE_FACTORY_ENTRY(SOURCEAPINAME, __VA_ARGS__)
 #define ENTRY_TEMPLATED(SOURCEAPINAME, ...)                                    \
   TEMPLATED_CALL_FACTORY_ENTRY(SOURCEAPINAME, __VA_ARGS__)
-#include "APINamesTexture.inc"
+#include "RulesLang/APINamesTexture.inc"
 #undef ENTRY_RENAMED
 #undef ENTRY_TEXTURE
 #undef ENTRY_UNSUPPORTED
