+++ clang/lib/DPCT/RulesLang/CallExprRewriterWarp.cpp
@@ -1,4 +1,4 @@
-//===--------------- CallExprRewriterComplex.cpp --------------------------===//
+//===--------------- CallExprRewriterWarp.cpp -----------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -12,11 +12,12 @@
 namespace clang {
 namespace dpct {
 
-void CallExprRewriterFactoryBase::initRewriterMapComplex() {
+void CallExprRewriterFactoryBase::initRewriterMapWarp() {
   RewriterMap->merge(
       std::unordered_map<std::string,
                          std::shared_ptr<CallExprRewriterFactoryBase>>({
-#include "APINamesComplex.inc"
+
+#include "RulesLang/APINamesWarp.inc"
       }));
 }
