+++ clang/lib/DPCT/RulesLang/Math/CallExprRewriterMath.h
