+++ clang/lib/DPCT/RulesLang/Math/CommonMacroDefinition.h
