+++ clang/lib/DPCT/RulesLang/Math/RewriterCXXAPIRoutines.cpp
