+++ clang/lib/DPCT/RulesLang/Math/RewriterDoublePrecisionIntrinsics.cpp
