+++ clang/lib/DPCT/RulesLang/Math/RewriterHalf2ArithmeticFunctions.cpp
