+++ clang/lib/DPCT/RulesLang/Math/RewriterHalf2ComparisonFunctions.cpp
