+++ clang/lib/DPCT/RulesLang/Math/RewriterHalf2MathFunctions.cpp
