+++ clang/lib/DPCT/RulesLang/Math/RewriterHalfArithmeticFunctions.cpp
