+++ clang/lib/DPCT/RulesLang/Math/RewriterHalfComparisonFunctions.cpp
