+++ clang/lib/DPCT/RulesLang/Math/RewriterHalfMathFunctions.cpp
