+++ clang/lib/DPCT/RulesLang/Math/RewriterIntegerIntrinsics.cpp
