+++ clang/lib/DPCT/RulesLang/Math/RewriterIntegerMathematicalFunctions.cpp
