+++ clang/lib/DPCT/RulesLang/Math/RewriterSIMDIntrinsics.cpp
