+++ clang/lib/DPCT/RulesLang/Math/RewriterSTDFunctions.cpp
