+++ clang/lib/DPCT/RulesLang/Math/RewriterSinglePrecisionIntrinsics.cpp
